feat: seperate svg files
