--- src/assets/svg/GoBackSvg.tsx
@@ -1,19 +1,0 @@
-import React from 'react';
-import { useRouter } from 'next/router';
-import { SvgIcon } from './SvgIcon';
-
-const GoBackSvg = () => {
-  const router = useRouter();
-
-  const handleClick = () => {
-    router.back();
-  };
-
-  return (
-    <SvgIcon viewBox='0 0 320 512' onClick={handleClick}>
-      <path d='M9.4 233.4c-12.5 12.5-12.5 32.8 0 45.3l192 192c12.5 12.5 32.8 12.5 45.3 0s12.5-32.8 0-45.3L77.3 256 246.6 86.6c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0l-192 192z' />
-    </SvgIcon>
-  );
-};
-
-export default GoBackSvg;
