+++ src/assets/svg/SvgStyle.tsx
@@ -1,12 +1,14 @@
 import styled from 'styled-components';
 
-export const SvgIcon = styled.svg`
-  display: flex;
-  justify-content: center;
-  margin: auto;
-  width: 24px;
-  height: 24px;
-  cursor: pointer;
+export const SvgIcon = styled.span`
+  svg {
+    display: flex;
+    justify-content: center;
+    margin: auto;
+    width: 24px;
+    height: 24px;
+    cursor: pointer;
+  }
 `;
 
 export const SvgWrapper = styled.div`
